feat: initially handle changing partOf
